backend/requirements.txt
# pip install -r requirements.txt

bcrypt
cryptography
fastapi
python-jose
python-multipart
python-dotenv
SQLAlchemy
uvicorn
passlib
jose
psycopg2
pyotp
Pillow